profile screen header + few tweaks to start setting up profile page improvements
